Added command line arg to test script.

+++ test.sh
@@ -3,4 +3,6 @@
 rm -rf Supergrammar-parsed-new.json
 mono giza/bin/Debug/giza.exe --super Supergrammar.txt > Supergrammar-parsed-new.json
 diff Supergrammar-parsed-ref.json Supergrammar-parsed-new.json
-rm -rf Supergrammar-parsed-new.json
+if [ "$1" == "" ]; then
+  rm -rf Supergrammar-parsed-new.json
+fi
